refactor: introduce shared api contracts
