+++ src/api/rooms.ts
@@ -1,40 +1,15 @@
 import { apiRequest } from "@/api/core";
+import type {
+  CreateRoomPayload,
+  RoomMessage,
+  RoomSummary,
+} from "@/shared/contracts/chat";
 
-export interface RoomSummary {
-  id: string;
-  name: string;
-  type?: "public" | "private" | "irc";
-  createdAt: number;
-  members?: string[];
-  userCount: number;
-  ircHost?: string;
-  ircPort?: number;
-  ircTls?: boolean;
-  ircChannel?: string;
-  ircServerLabel?: string;
-}
-
-export interface RoomMessage {
-  id: string;
-  roomId: string;
-  username: string;
-  content: string;
-  timestamp: number;
-  clientId?: string;
-}
-
-export interface CreateRoomPayload {
-  type: "public" | "private" | "irc";
-  name?: string;
-  members?: string[];
-  /** Registered server id from GET /api/irc/servers (required for non-admin IRC rooms). */
-  ircServerId?: string;
-  ircHost?: string;
-  ircPort?: number;
-  ircTls?: boolean;
-  ircChannel?: string;
-  ircServerLabel?: string;
-}
+export type {
+  CreateRoomPayload,
+  RoomMessage,
+  RoomSummary,
+} from "@/shared/contracts/chat";
 
 export async function listRooms(): Promise<{ rooms: RoomSummary[] }> {
   return apiRequest<{ rooms: RoomSummary[] }>({
